# fix slow compute metadata on PNGs due to getexif on png