@swheaton 2025-03-10
Here's the original PR introducing EXIF stuff: https://github.com/voxel51/fiftyone/pull/3824